Remove desktop which is incompatible with our license
I wish I had noticed this sooner. When I had taken over this plugin
from the original author, it did not have a license assigned to it, but
I did not properly audit every bit of the code. Desktop has an LGPL
license while we use MIT. This is fairly easy to fix for us as desktop
was only used to open a web browser, and really, only for Linux.
I've personally always just used `xdg` to open browsers on Linux, and
that is what we'll do here. If there is some case we are missing, we'll
solve it as we discover it, but we should not have been and cannot
use desktop with our license.